### 2024-01-24 #6275 @editorialbot
[REVIEW]: PyHD - PyHeatDemand - Processing Tool for Heat Demand Data
**Submitting author:** @AlexanderJuestel (<a href="http://orcid.org/0000-0003-0980-7479">Alexander Jüstel</a>)
**Repository:** https://github.com/AlexanderJuestel/pyheatdemand
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
