### 2024-01-24 #6275 @AlexanderJuestel
> ```
>
> MISSING DOIs
>
> - 10.20944/preprints202312.1795.v1 may be a valid DOI for title: Unveiling the spatial distribution of heat demand in North-West-Europe compiled with national heat Consumption data

> ```

`10.20944/preprints202312.1795.v1` is the preprint version of the peer-reviewed article at `10.3390/en17020481` which was first first-authored by me and applies the PyHeatDemand tool to create a transnational-harmonized heat demand map of the Interreg North-West-Europe Region
