+++ pyheatdemand/processing.py
@@ -35,13 +35,18 @@ def create_polygon_mask(gdf: Union[gpd.GeoDataFrame, Polygon],
             Size of the rasterized squares in meters, e.g. ``step_size=100``.
         crop_gdf : bool, default: ``False``
             Boolean to either crop the GeoDataFrame to the outline or return it as is, e.g. ``crop_gdf=False``.
-        crs : Union[str, pyproj.crs.crs.CRS]
+        crs : Union[str, pyproj.crs.crs.CRS], default: ``None``
             Coordinate Reference System when providing Shapely Polygons as input.
 
     Returns
     --------
         gdf_mask : gpd.GeoDataFrame
             GeoDataFrame containing the masked polygons.
+
+            ========== ===============================
+            Index      Index of each mask polygon
+            geometry   Geometry of each mask polygon
+            ========== ===============================
 
     Raises
     ______
@@ -53,10 +58,17 @@ def create_polygon_mask(gdf: Union[gpd.GeoDataFrame, Polygon],
 
         >>> mask = create_polygon_mask(gdf=gdf, step_size=100, crop_gdf=True)
         >>> mask
-            geometry
-        0   POLYGON ((2651470.877 2135999.353, 2661470.877...
-        1   POLYGON ((2651470.877 2145999.353, 2661470.877...
-        2   POLYGON ((2651470.877 2155999.353, 2661470.877...
+
+        =======  ===================================================
+        Index    geometry
+        =======  ===================================================
+        0        POLYGON ((2651470.877 2135999.353, 2661470.877...
+        1        POLYGON ((2651470.877 2145999.353, 2661470.877...
+        2        POLYGON ((2651470.877 2155999.353, 2661470.877...
+        =======  ===================================================
+
+    See Also
+    ________
 
 
     """
@@ -131,7 +143,12 @@ def refine_mask(mask: gpd.GeoDataFrame,
     Returns
     _______
         grid_refs : gpd.GeoDataFrame
-            GeoDataFrame containing the refined mask polygons.
+            GeoDataFrame containing the refined mask polygons. Data columns are as follows:
+
+            ========== ===============================
+            Index      Index of each mask polygon
+            geometry   Geometry of each mask polygon
+            ========== ===============================
 
     Raises
     ______
@@ -145,11 +162,14 @@ def refine_mask(mask: gpd.GeoDataFrame,
 
         >>> mask = refine_mask(mask=mask, data=data, num_of_points=100, cell_size=1000)
         >>> mask
-            geometry
-        0   POLYGON ((2651470.877 2135999.353, 2661470.877...
-        1   POLYGON ((2651470.877 2145999.353, 2661470.877...
-        2   POLYGON ((2651470.877 2155999.353, 2661470.877...
 
+        =======  ===================================================
+        Index    geometry
+        =======  ===================================================
+        0        POLYGON ((2651470.877 2135999.353, 2661470.877...
+        1        POLYGON ((2651470.877 2145999.353, 2661470.877...
+        2        POLYGON ((2651470.877 2155999.353, 2661470.877...
+        =======  ===================================================
 
     """
     # Checking that the mask is of type GeoDataFrame
@@ -229,7 +249,12 @@ def quad_tree_mask_refinement(mask: gpd.GeoDataFrame,
     Returns
     _______
         grid_refs : gpd.GeoDataFrame
-            GeoDataFrame containing the refined mask polygons.
+            GeoDataFrame containing the refined mask polygons. Data columns are as follows:
+
+            ========== ===============================
+            Index      Index of each mask polygon
+            geometry   Geometry of each mask polygon
+            ========== ===============================
 
     Raises
     ______
@@ -243,10 +268,14 @@ def quad_tree_mask_refinement(mask: gpd.GeoDataFrame,
 
[…]
