### 2024-01-27 #6275 @AlexanderJuestel
FYI: I tried to update the API Reference and documentation today but RTD failed to build due to an error with Sphinx versions and dependencies. Will have to wait for a fix here. See also the issue reported here: https://github.com/sphinx-doc/sphinx/issues/11890

https://readthedocs.org/projects/pyhd/builds/23253592/
