### 2024-02-12 #6275 @nmstreethran
@timtroendle I just finished reviewing the documentation and installation instructions (see https://github.com/AlexanderJuestel/pyheatdemand/issues/28). I'm still going through the paper and examples to verify the functionality. I expect to complete my review by the end of the week.
