### 2024-02-13 AlexanderJuestel/pyheatdemand#28 @AlexanderJuestel
@nmstreethran, wow thank you for this thorough review. I will go through your remarks and implement them in the next coming days!
