### 2024-02-17 AlexanderJuestel/pyheatdemand#28 @AlexanderJuestel
>
> ## Documentation
> * For consistency, you could structure the installation instructions in the README to match that of the readthedocs documentation, i.e. split Pip and Conda installation instructions as subsections.
> ```

Docs and Readme were changed in https://github.com/AlexanderJuestel/pyheatdemand/commit/5e0462cec41ee0c66f578c742fdb8159b3f2dfd4

>
> * Feature the documentation more prominently in the README. Right now, there's a badge for the docs build but a section with a link to the documentation on readthedocs will be useful. You can maybe merge the API reference / local documentation build instructions here.
> ```

A reference to the documentation page was added to the readme in https://github.com/AlexanderJuestel/pyheatdemand/commit/a5d31f6d7b498f1e673cb767b3451e1bfa4c0021.

>
> * Add more details on installing from source, i.e. using `environment_dev.yml` and `requirements.txt` after cloning the repository.
> * Add a link to the contributing guidelines to the README.
> * Although you have included instructions for continuous integration, I think you should add the testing instructions to the documentation, perhaps under the installation instructions, for testing locally.
> * `pytest-cov` isn't included in `environment_dev.yml` and `requirements.txt`, but I think it should be for the purpose of testing locally.
> * I also noted that the packages in `environment_dev.yml` and `requirements.txt` do not match. Perhaps you could include the missing sphinx packages in `requirements.txt` for consistency?
>

> ```
