### 2024-02-17 AlexanderJuestel/pyheatdemand#28 @AlexanderJuestel
@nmstreethran, please see my edits in the linked commits above. I should have resolved all your requested changes. Thanks again for this thorough review so far.
