### 2024-02-17 #6275 @AlexanderJuestel
> @timtroendle I just finished reviewing the documentation and installation instructions (see [AlexanderJuestel/pyheatdemand#28](https://github.com/AlexanderJuestel/pyheatdemand/issues/28)). I'm still going through the paper and examples to verify the functionality. I expect to complete my review by the end of the week.

@timtroendle I have implemented all the requested changes by @nmstreethran on this issue :)
