### 2024-02-19 AlexanderJuestel/pyheatdemand#28 @nmstreethran
@AlexanderJuestel That's great, thank you for addressing my comments. I'll get back to you tomorrow with the rest of my comments on the paper and examples.
