### 2024-02-20 AlexanderJuestel/pyheatdemand#29 @nmstreethran
[JOSS Review] Paper
@AlexanderJuestel here are some initial comments on the paper. I have to do a quick check again so I'll comment if I have any further feedback.

https://github.com/openjournals/joss-reviews/issues/6275

## General comments

I'm curious how much lower the 100 m x 100 m resolution is compared to that of a building, since the paper states the following:

> Evaluating the heat demand (usually in MWh = Mega Watt Hours) on a national or regional scale, including space and water heating for each apartment or each
> building for every day of a year separately is from a perspective of resolution (spatial and temporal scale) and computing power
> not feasible. Therefore, heat demand maps summarize the heat demand on a lower spatial resolution (e.g. 100 m x 100 m
> raster) cumulated for one year (lower temporal resolution) for different sectors such as the residential and tertiary
> sectors.

MWh was also defined here but it has already been used in the summary twice. I suggest moving the definition to the first instance where MWh is used. You can perhaps define it in the form of a footnote.

The acronym "HD" is used but is not defined anywhere in the paper.

According to my checklist, the statement of need must include the target audience of PyHeatDemand. The target audience should also be included in the documentation.

## Formatting

I found some minor formatting issues in the [PDF article](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06275/joss.06275/10.21105.joss.06275.pdf) and docs which I've listed below.

The blank lines in the following code are breaking the sentence, so they should be removed.

https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/docs/source/about.rst?plain=1#L6-L8

https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/joss/paper.md?plain=1#L26-L28

https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/joss/paper.md?plain=1#L35-L37

The table isn't rendered properly. I think you need to remove all of these dashes, except for the one below the header row, to fix it. Also, you have referenced the table as Tab. 1. Since the table doesn't have a caption, you can either add a caption, or remove the reference to Tab. 1 and just refer to it as the "table below".

https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/joss/paper.md?plain=1#L93-L101

## Paper structure

I think the State of the field, PyHeatDemand Outlook, and PyHeatDemand Resources sections could possibly be merged with the previous sections as they are short. The outlook could go in the Summary. The state of the field should probably be in the statement of need. You can remove the GitHub repository and documentation link as the PDF of the paper already has a link to the repo. The DGE Rollout Webviewer and a reference to Herbst, 2021 could go in the summary as well, where you mention in the final sentence that the package was developed during the DGE Rollout project. You can also move Jüstel et al., 2023 here as an application of the package.

## References

There are no citations provided in the summary and statement of need. I suggest adding citations to the following:

- Summary
    - the purpose of heat demand quantification and mapping
- Statement of need:
    - energy consumption, primary energy, and thermal energy statistics
    - EU energy efficiency directive and emission reduction target
    - example of heat demand input values for the residential and commercial sectors that are easily accessible and assessable

## Bibliography

The DOI for Jüstel et al., 2023 is missing in the PDF. Change `note=` to `doi=` to fix this:

https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/joss/paper.bib#L220

Herbst, 2021 seems like it's missing several authors?

Some of the titles are not capitalised properly in the PDF output of the paper. These words/titles (emphasised in bold) should be enclosed in braces in the bib file to preserve the capitalisation:

- Herbst, K., E. (2021). A heat demand map of **north-west europe**
- Esmukov, K., & others. (2023). GeoPy: Geocoding library for **python**
- Gillies, S., & others. (2013). Rasterio: Geospatial raster **i/o** for Python programmers.
- Meha, D., Novosel, T., & Duić, N. (2020). Bottom-up and top-down heat demand mapping methods for small municipalities, case **gllogoc**
