### 2024-02-20 AlexanderJuestel/pyheatdemand#30 @nmstreethran
[JOSS Review] Code and examples
@AlexanderJuestel here are my comments on the code and examples.

https://github.com/openjournals/joss-reviews/issues/6275

## Code

I recommend linting `processing.py` for some suggestions on improving the code style. Here are three suggested changes:

- change the following to `shapes = list(zip(gdf_hd['geometry'], gdf_hd['HD']))`
  https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/pyheatdemand/processing.py#L782

- bare except should be avoided if possible
  https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/pyheatdemand/processing.py#L963-L964

- these should use `if a not in b` instead of `if not a in b`
  https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/pyheatdemand/processing.py#L513
  https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/pyheatdemand/processing.py#L656
  https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/pyheatdemand/processing.py#L1372
  https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/pyheatdemand/processing.py#L1441

Additionally, these two snippets seem to be identical. Is it possible to define it as a function for use in these two places to avoid code repetition?

https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/pyheatdemand/processing.py#L495-L532

https://github.com/AlexanderJuestel/pyheatdemand/blob/73f41fd08b29771bd81dfc81e9801915cb1c8e94/pyheatdemand/processing.py#L638-L675

## Example notebooks

To import the library from the cloned repository, using `from pyheatdemand import processing` was sufficient. Using `sys` was not necessary (also did not work for me; I'm not using Windows though). In the example notebooks, the `sys` path uses both `pyhd` and `pyheatdemand`; all instances of `pyhd` should be fixed.

The data paths used in the notebooks are inconsistent, e.g. `../../../data/`, `../data/`, etc. I noticed that many of the datasets used are available in the `test` folder. Could you make use of these in the example notebooks? This will allow potential users to execute the notebooks themselves and reproduce the results.

In some notebooks (for example, `02_Processing_Data_Type_I_Raster.ipynb`), the data was reprojected to `epsg:3034`. However, subsequent code uses `to_crs('EPSG:3034')`, which is redundant and can be removed.

The semicolon in `show();` should be removed.

In the notebook `14_Refining_Polygon_Mask.ipynb`, under 'Refining the mask', there's a cell with a `TypeError` output that should be deleted.

Some other suggestions for improvement are to include axis labels / titles / legends in as many plots as possible (where appropriate) as there are a lot of images in the notebooks and this will help with some context. For example, the histograms could use titles or axis labels, and the heat demand colour bars could use a legend title.
