### 2024-02-20 #6275 @nmstreethran
Sorry, I was only halfway through my review. I've opened two more issues now and will complete my checklist once these are resolved.
https://github.com/AlexanderJuestel/pyheatdemand/issues/29
https://github.com/AlexanderJuestel/pyheatdemand/issues/30
