+++ pyheatdemand/processing.py
@@ -441,6 +441,68 @@ def create_outline(gdf: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
                             crs=gdf.crs)
 
 
+def _check_hd_input(hd_gdf: gpd.GeoDataFrame,
+                    mask_gdf: Union[gpd.GeoDataFrame, Polygon],
+                    hd_data_column: str = ''):
+    """Check heat demand input data.
+
+    Parameters
+    __________
+        hd_gdf : gpd.GeoDataFrame
+            Heat demand data as GeoDataFrame.
+        mask_gdf : Union[gpd.GeoDataFrame, shapely.geometry.Polygon]
+            Mask for the output Heat Demand Data.
+        hd_data_column : str, default: ``''``
+            Name of the column that contains the Heat Demand Data, e.g. ``hd_data_column='HD'``.
+
+    Raises
+    ______
+        TypeError
+            If the wrong input data types are provided.
+
+    """
+    # Converting Shapely Polygon to GeoDataFrame
+    if isinstance(mask_gdf, Polygon):
+        mask_gdf = gpd.GeoDataFrame(geometry=[mask_gdf],
+                                    crs=hd_gdf.crs)
+
+    # Checking that the hd_gdf is of type GeoDataFrame
+    if not isinstance(hd_gdf, gpd.GeoDataFrame):
+        raise TypeError('The heat demand gdf must be provided as GeoDataFrame')
+
+    # Checking that the mask_gdf is of type GeoDataFrame
+    if not isinstance(mask_gdf, gpd.GeoDataFrame):
+        raise TypeError('The mask gdf must be provided as GeoDataFrame')
+
+    # Checking that the Heat Demand Data Column is provided as string
+    if not isinstance(hd_data_column, str):
+        raise TypeError('The heat demand data column must be provided as string')
+
+    # Checking that the HD Data Column is in the HD GeoDataFrame
+    if hd_data_column not in hd_gdf:
+        raise ValueError('%s is not a column in the GeoDataFrame' % hd_data_column)
+
+    # Reprojecting Data if necessary
+    if mask_gdf.crs != hd_gdf.crs:
+        hd_gdf = hd_gdf.to_crs(mask_gdf.crs)
+
+    # Exploding MultiPolygons
+    if any(shapely.get_type_id(hd_gdf.geometry) == 6):
+        hd_gdf = hd_gdf.explode(index_parts=True).reset_index(drop=True)
+
+    # Assigning area to Polygons
+    if all(shapely.get_type_id(hd_gdf.geometry) == 3):
+        # Assigning area of original geometries to GeoDataFrame
+        hd_gdf['area'] = hd_gdf.area
+
+    # Assigning lengths to LineStrings
+    elif all(shapely.get_type_id(hd_gdf.geometry) == 1):
+        # Assigning length of original geometries to GeoDataFrame
+        hd_gdf['length'] = hd_gdf.length
+
+    return hd_gdf, mask_gdf, hd_data_column
+
+
 def calculate_hd(hd_gdf: gpd.GeoDataFrame,
                  mask_gdf: Union[gpd.GeoDataFrame, Polygon],
                  hd_data_column: str = '') -> gpd.GeoDataFrame:
@@ -492,44 +554,10 @@ def calculate_hd(hd_gdf: gpd.GeoDataFrame,
         calculate_hd_sindex : Calculate Heat Demand using Spatial Indices.
 
     """
-    # Converting Shapely Polygon to GeoDataFrame
-    if isinstance(mask_gdf, Polygon):
-        mask_gdf = gpd.GeoDataFrame(geometry=[mask_gdf],
[…]
