### 2024-02-20 AlexanderJuestel/pyheatdemand#30 @AlexanderJuestel
> # Code

All requested Code changes were implemented in https://github.com/AlexanderJuestel/pyheatdemand/commit/50110d9c1ee7b7d3aa0d9968d325bc50d8ef8bb6
