### 2024-02-20 AlexanderJuestel/pyheatdemand#30 @AlexanderJuestel
> ## Example notebooks
> To import the library from the cloned repository, using `from pyheatdemand import processing` was sufficient. Using `sys` was not necessary (also did not work for me; I'm not using Windows though). In the example notebooks, the `sys` path uses both `pyhd` and `pyheatdemand`; all instances of `pyhd` should be fixed.
>

Imports were fixed in https://github.com/AlexanderJuestel/pyheatdemand/commit/543ae7808db5c59fbe618b2b1c12c3b15d7f0368

> The data paths used in the notebooks are inconsistent, e.g. `../../../data/`, `../data/`, etc. I noticed that many of the datasets used are available in the `test` folder. Could you make use of these in the example notebooks? This will allow potential users to execute the notebooks themselves and reproduce the results.
>
> In some notebooks (for example, `02_Processing_Data_Type_I_Raster.ipynb`), the data was reprojected to `epsg:3034`. However, subsequent code uses `to_crs('EPSG:3034')`, which is redundant and can be removed.
>
> The semicolon in `show();` should be removed.
>
> In the notebook `14_Refining_Polygon_Mask.ipynb`, under 'Refining the mask', there's a cell with a `TypeError` output that should be deleted.
>
> Some other suggestions for improvement are to include axis labels / titles / legends in as many plots as possible (where appropriate) as there are a lot of images in the notebooks and this will help with some context. For example, the histograms could use titles or axis labels, and the heat demand colour bars could use a legend title.
