### 2024-02-21 4bdc771
Revert "FixNotebooksandData"
This reverts commit 05e1d1d0
