+++ test/data/HD_Aachen.qmd
@@ -1,27 +1,0 @@
-<!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
-<qgis version="3.32.2-Lima">
-  <identifier></identifier>
-  <parentidentifier></parentidentifier>
-  <language></language>
-  <type>dataset</type>
-  <title></title>
-  <abstract></abstract>
-  <links/>
-  <dates/>
-  <fees></fees>
-  <encoding></encoding>
-  <crs>
-    <spatialrefsys nativeFormat="Wkt">
-      <wkt></wkt>
-      <proj4>+proj=lcc +lat_0=52 +lon_0=10 +lat_1=35 +lat_2=65 +x_0=4000000 +y_0=2800000 +ellps=GRS80 +units=m +no_defs</proj4>
-      <srsid>0</srsid>
-      <srid>0</srid>
-      <authid></authid>
-      <description></description>
-      <projectionacronym></projectionacronym>
-      <ellipsoidacronym></ellipsoidacronym>
-      <geographicflag>false</geographicflag>
-    </spatialrefsys>
-  </crs>
-  <extent/>
-</qgis>
