### 2024-02-21 AlexanderJuestel/pyheatdemand#30 @AlexanderJuestel
@nmstreethran I think I am done with your comments here. Feel free to have a look and thanks for bringing up some important points.
