+++ joss/paper.md
@@ -23,112 +23,60 @@ bibliography: paper.bib
 ---
 
 # Summary
-**PyHeatDemand** is an open-source Python package for processing and harmonizing multi-scale-multi-type heat demand input data for
-constructing local to transnational harmonized heat demand maps (rasters). Knowledge about the heat demand in megawatt hours per area and per year of a respective building, 
-district, city, state, country, or even on a continental scale is crucial for an adequate heat demand analysis or 
-planning for providing power plant capacities. Mapping of the heat demand may also identify potential areas for new
-district heating networks or even geothermal power plants for climate-friendly heat production. 
+**PyHeatDemand** is an open-source Python package for processing and harmonizing multi-scale-multi-type heat demand input data for constructing local to transnational harmonized heat demand maps (rasters). Knowledge about the heat demand in megawatt hours per area and per year of a respective building, district, city, state, country, or even on a continental scale is crucial for an adequate heat demand analysis or planning for providing power plant capacities. Mapping of the heat demand may also identify potential areas for new district heating networks or even geothermal power plants for climate-friendly heat production. 
 
-The aim of **PyHeatDemand** is to provide processing tools for heat demand input data of various categories on various scales. This
-includes heat demand input data provided as rasters or gridded polygons, heat demand input data associated with administrative areas
-(points or polygons), with building footprints (polygons), with street segments (lines), or with addresses directly provided in 
-kWh or MWh but also as gas usage, district heating usage, or other sources of heat. It is also possible to calculate the heat demand
-based on a set of cultural data sets (building footprints, height of the buildings, population density, building type, etc.).
-The study area is first divided into a coarse
-mask before heat demands are calculated and harmonized for each cell with the size of the target resolution (e.g. 100 m
-x 100 m for states). We hereby make use of different spatial operations implemented in the GeoPandas and Shapely
-packages. The final heat demand map will be created utilizing the Rasterio package. Next to processing tools for the heat demand input data, workflows for analyzing the final heat demand map through
-the Rasterstats package are provided. 
+The aim of **PyHeatDemand** is to provide processing tools for heat demand input data of various categories on various scales. This includes heat demand input data provided as rasters or gridded polygons, heat demand input data associated with administrative areas (points or polygons), with building footprints (polygons), with street segments (lines), or with addresses directly provided in kWh or MWh but also as gas usage, district heating usage, or other sources of heat. It is also possible to calculate the heat demand based on a set of cultural data sets (building footprints, height of the buildings, population density, building type, etc.). The study area is first divided into a coarse mask before heat demands are calculated and harmonized for each cell with the size of the target resolution (e.g. 100 m x 100 m for states). We hereby make use of different spatial operations implemented in the GeoPandas and Shapely packages. The final heat demand map will be created utilizing the Rasterio package. Next to processing tools for the heat demand input data, workflows for analyzing the final heat demand map through the Rasterstats package are provided. 
 
 **PyHeatDemand** was developed as a result of works carried out within the Interreg NWE project DGE Rollout (Rollout of Deep Geothermal Energy).
 
 
 # Statement of need 
-Space and water heating for residential and commercial buildings amount to a third of the European Union’s total final 
-energy consumption. Approximately 75% of the primary energy and 50% of the thermal energy are still produced by burning 
-fossil fuels, leading to high greenhouse gas emissions in the heating sector. The transition from centralized 
-fossil-fueled district heating systems such as coal or gas power plants to district heating systems sourced by renewable
-energies such as geothermal energy or more decentralized individual solutions for city districts makes it necessary to 
-map the heat demand for a more accurate planning of power plant capacities. In addition, heating and cooling plans 
-become necessary according to directives of the European Union regarding energy efficiency to reach its aim of reducing 
-greenhouse gas emissions by 55% of the 1990-levels by 2030. 
[…]
