+++ docs/source/about.rst
@@ -3,27 +3,11 @@
 About
 =====
 
-**PyHeatDemand** is an open-source Python package for processing and harmonizing multi-scale-multi-type heat demand input data for
+**PyHeatDemand** is an open-source Python package for processing and harmonizing multi-scale-multi-type heat demand input data for constructing local to transnational harmonized heat demand maps (rasters). Knowledge about the heat demand (MWh/area/year) of a respective building, district, city, state, country, or even on a continental scale is crucial for an adequate heat demand analysis or planning for providing power plant capacities. Mapping of the heat demand may also identify potential areas for new district heating networks or even geothermal power plants for climate-friendly heat production.
 
-constructing local to transnational harmonized heat demand maps (rasters). Knowledge about the heat demand (MWh/area/year) of a respective building,
-district, city, state, country, or even on a continental scale is crucial for an adequate heat demand analysis or
-planning for providing power plant capacities. Mapping of the heat demand may also identify potential areas for new
-district heating networks or even geothermal power plants for climate-friendly heat production.
+The aim of **PyHeatDemand** is to provide processing tools for heat demand input data of various categories on various scales. This includes heat demand input data provided as rasters or gridded polygons, heat demand input data associated with administrative areas (points or polygons), with building footprints (polygons), with street segments (lines), or with addresses directly provided in MWh but also as gas usage, district heating usage, or sources of heat. It is also possible to calculate the heat demand based on a set of cultural data sets (building footprints, height of the buildings, population density, building type, etc.). The study area is first divided into a coarse mask before heat demands are calculated and harmonized for each cell with the size of the target resolution (e.g. 100 m x 100 m for states). We hereby make use of different spatial operations implemented in the `GeoPandas <https://geopandas.org/en/stable/>`_ and `Shapely <https://shapely.readthedocs.io/en/stable/manual.html>`_ packages. The final heat demand map will be created utilizing the `Rasterio <https://rasterio.readthedocs.io/en/stable/>`_ package. Next to processing tools for the heat demand input data, workflows for analyzing the final heat demand map through the `Rasterstats <https://pythonhosted.org/rasterstats/>`_ package are provided.
 
-The aim of **PyHeatDemand** is to provide processing tools for heat demand input data of various categories on various scales. This
-includes heat demand input data provided as rasters or gridded polygons, heat demand input data associated with administrative areas
-(points or polygons), with building footprints (polygons), with street segments (lines), or with addresses directly provided in
-MWh but also as gas usage, district heating usage, or sources of heat. It is also possible to calculate the heat demand
-based on a set of cultural data sets (building footprints, height of the buildings, population density, building type, etc.).
-The study area is first divided into a coarse
-mask before heat demands are calculated and harmonized for each cell with the size of the target resolution (e.g. 100 m
-x 100 m for states). We hereby make use of different spatial operations implemented in the `GeoPandas <https://geopandas.org/en/stable/>`_ and `Shapely <https://shapely.readthedocs.io/en/stable/manual.html>`_
-packages. The final heat demand map will be created utilizing the `Rasterio <https://rasterio.readthedocs.io/en/stable/>`_ package. Next to processing tools for the heat demand input data, workflows for analyzing the final heat demand map through
-the `Rasterstats <https://pythonhosted.org/rasterstats/>`_ package are provided.
-
[…]
