+++ joss/paper.md
@@ -41,9 +41,10 @@ Evaluating the annual heat demand (HD, usually in MWh = megawatt Hours) on a nat
 
 Heat demand maps can be calculated using either a top-down approach or a bottom-up approach (Fig. \ref{fig0}). For the top-down approach, aggregated heat demand input data for a certain area will be distributed according to higher resolution data sets (e.g. population density, landuse, etc.). In contrast to that, the bottom-up approach allows aggregating heat demand of higher resolution data sets to a lower resolution (e.g. from building level to a 100 m x 100 m raster).
 
+**PyHeatDemand** processes geospatial data such as vector data (points, lines, polygons), raster data or address data. Therefore, we make use of the functionality implemented in well-known geospatial packages such as GeoPandas [@geopandas], Rasterio [@rasterio], Rasterstats [@rasterstats], GeoPy [@geopy], or OSMnx [@osmnx] and their underlying dependencies such as Shapely [@shapely], Pandas [@pandas], or NumPy [@numpy]. In particular, we are utilizing the powerful implementation of [Spatial Indices](https://geopandas.org/en/stable/docs/reference/sindex.html) in GeoPandas allowing for processing speed-ups by orders of magnitudes compared to performing regular overlays and spatial joins for the processing of heat demand input data.
+
 ![Input and output data for top-down and bottom-up approaches. Note, that the resulting spatial resolution can be the same for both approaches, but the spatial value of information is usually lower using a top-down approach. \label{fig0}](../docs/images/fig0.png)
 
-**PyHeatDemand** processes geospatial data such as vector data (points, lines, polygons), raster data or address data. Therefore, we make use of the functionality implemented in well-known geospatial packages such as GeoPandas [@geopandas], Rasterio [@rasterio], Rasterstats [@rasterstats], GeoPy [@geopy], or OSMnx [@osmnx] and their underlying dependencies such as Shapely [@shapely], Pandas [@pandas], or NumPy [@numpy]. In particular, we are utilizing the powerful implementation of [Spatial Indices](https://geopandas.org/en/stable/docs/reference/sindex.html) in GeoPandas allowing for processing speed-ups by orders of magnitudes compared to performing regular overlays and spatial joins for the processing of heat demand input data.
 
 The creation of a heat demand map follows a general workflow (Fig. \ref{fig1}) followed by a data-category-specific workflow for five defined input data categories (Fig. \ref{fig2} \& \ref{fig3}). The different input data categories are listed in the table below. 
 
