### 2024-02-21 AlexanderJuestel/pyheatdemand#29 @AlexanderJuestel
> # Formatting

The formatting issues were addressed in https://github.com/AlexanderJuestel/pyheatdemand/commit/becfecbe3141e46c79c3350e322c1df4e0571529

All line breaks were removed in the paper and the about section. The Table was fixed.

![image](https://github.com/AlexanderJuestel/pyheatdemand/assets/45469915/0d9a6583-2d24-4783-bb46-7f9af4ce595d)
