### 2024-02-21 AlexanderJuestel/pyheatdemand#29 @AlexanderJuestel
> Paper Structure

The paper structure was addressed in https://github.com/AlexanderJuestel/pyheatdemand/commit/6e4b83d2cfa6a8d81745c1ebe8c052b9fc71ead9

The State of the field was moved to the statement of need section. The outlook was split. The first part went into the summary and the second part was merged with the Resources. This is the only of the three sections that remains now.
