+++ joss/paper.bib
@@ -14,7 +14,7 @@
 
 @article{gempy,
   author = {Miguel de la Varga and Alexander Schaaf and Florian Wellmann},
-  title = {GemPy 1.0: open-source stochastic geological modeling and inversion},
+  title = {{GemPy} 1.0: open-source stochastic geological modeling and inversion},
   pages = {1-32},
   year = {2019},
   journal = {Geoscientific Model Development}
@@ -23,7 +23,7 @@
 @software{rasterio,
   author =    {Sean Gillies and others},
   organization = {Mapbox},
-  title =     {Rasterio: geospatial raster I/O for {Python} programmers},
+  title =     {Rasterio: geospatial raster {I/O} for {Python} programmers},
   year =      {2013},
   url = "https://github.com/mapbox/rasterio"
 }
@@ -66,7 +66,7 @@
   number = {37},
   pages = {1450},
   author = {C. Bane Sullivan and Alexander Kaszynski},
-  title = {{PyVista}: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit ({VTK})},
+  title = {{PyVista}: {3D} plotting and mesh analysis through a streamlined interface for the {V}isualization {T}oolkit ({VTK})},
   journal = {Journal of Open Source Software}
 }
 
@@ -104,29 +104,16 @@
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
 
-@article{Map2Loop,
-author = {Jessell, M. and Ogarko, V. and de Rose, Y. and Lindsay, M. and Joshi, R. and Piechocka, A. and Grose, L. and de la Varga, M. and Ailleres, L. and Pirot, G.},
-title = {Automated geological map deconstruction for 3D model construction using
-\textit{map2loop} 1.0 and \textit{map2model} 1.0},
-journal = {Geoscientific Model Development},
-volume = {14},
-year = {2021},
-number = {8},
-pages = {5063-5092},
-url = {https://gmd.copernicus.org/articles/14/5063/2021/},
-doi = {10.5194/gmd-14-5063-2021}
-}
-
 @software{geopy,
   author =    {Kostya Esmukov and others},
   organization = {GeoPy},
-  title =     {GeoPy: Geocoding library for Python},
+  title =     {{GeoPy}: {G}eocoding library for {P}ython},
   year =      {2023},
   url = "https://github.com/geopy/geopy"
 }
 
 @article{osmnx,
-title = {OSMnx: New methods for acquiring, constructing, analyzing, and visualizing complex street networks},
+title = {{OSMnx}: New methods for acquiring, constructing, analyzing, and visualizing complex street networks},
 journal = {Computers, Environment and Urban Systems},
 volume = {65},
 pages = {126-139},
@@ -155,7 +142,7 @@ author = {T. Novosel and T. Pukšec and N. Duić and J. Domac},
 }
 
 @article{meha,
-title = {Bottom-up and top-down heat demand mapping methods for small municipalities, case Gllogoc},
+title = {Bottom-up and top-down heat demand mapping methods for small municipalities, case {G}llogoc},
 journal = {Energy},
 volume = {199},
 pages = {117429},
@@ -167,20 +154,20 @@ author = {Drilon Meha and Tomislav Novosel and Neven Duić},
 @software{rasterstats,
   author =    {Matthew Perry},
   organization = {Rasterstats},
-  title =     {Rasterstats: Summary statistics of geospatial raster datasets based on vector geometries.},
+  title =     {Rasterstats: {S}ummary statistics of geospatial raster datasets based on vector geometries.},
   year =      {2023},
[…]
