+++ joss/paper.md
@@ -78,11 +78,12 @@ The data processing for data categories 1 and 2 are very similar (Fig. \ref{fig2
 
 The data processing for data category 3 corresponds to a top-down approach (Fig. \ref{fig3} bottom). The heat demand represented as points for an administrative unit will be distributed across the area using higher-resolution data sets. In the case illustrated below, the distribution of Hotmaps data [@hotmaps] is used to distribute the available heat demands for the given administrative areas. For each administrative area, the provided total heat demand will distributed according to the share of each Hotmaps cell compared to the total Hotmaps heat demand of the respective area. The provided heat demand is now distributed across the cells and will treated from now on as category 1 or 2 input data to calculate the final heat demand map.  
 
-![The main steps of the methodology to process the provided HD polygons for the heat demand data category 2 (top) and category 3 (bottom). \label{fig3}](../docs/images/fig3.png)
-
 The data processing for data category 4 corresponds to a bottom-up approach. Here, the addresses will be converted using the GeoPy geolocator to coordinates. Based on these, the building footprints are extracted from OpenStreet Maps using OSMnx. From there on, the data will be treated as data category 2.
 
 If no heat demand input data is available, the heat demand can be estimated using cultural data such as population density, landuse, and building-specific heat usage [@novosel; @meha] which will be implemented in a later development stage.
+
+![The main steps of the methodology to process the provided HD polygons for the heat demand data category 2 (top) and category 3 (bottom). \label{fig3}](../docs/images/fig3.png)
+
 
 ## Processing Heat Demand Map Data
 
