### 2024-02-21 AlexanderJuestel/pyheatdemand#29 @AlexanderJuestel
> Bibliography

The bibliography issues were addressed in https://github.com/AlexanderJuestel/pyheatdemand/commit/f9e661b2aba217856a48705755604410117a3df7

There were indeed some brackets missing.
