+++ joss/paper.bib
@@ -206,3 +206,17 @@ author = {Finkel, R.A. and Bentley, J.L. },
     pages = {481},
     doi={10.3390/en17020481},
 }
+
+@Misc{Reiter,
+  author =    {Reiter, U. and Catenazzi, G. and Jakob, M. and Naegeli, C.},
+  title =     {Mapping and Analyses of the Current and Future (2020–2030) Heating/Cooling Fuel Deployment (Fossil/Renewables)—Work Package 1: Final Energy Consumption for the Year 2012},
+  year =      {2016},
+  url = "https://energy.ec.europa.eu/publications/mapping-and-analyses-current-and-future-2020-2030-heatingcooling-fuel-deployment-fossilrenewables_en"
+}
+
+@Misc{EnergyDirective,
+  author =    {{European Parliament and the Council}},
+  title =     {DIRECTIVE (EU) 2023/1791 OF THE EUROPEAN PARLIAMENT AND OF THE COUNCIL of 13 September 2023 on energy efficiency and amending Regulation (EU) 2023/955 (recast)},
+  year =      {2023},
+  url = "https://eur-lex.europa.eu/legal-content/EN/TXT/PDF/?uri=CELEX:32023L1791"
+}
