### 2024-02-21 AlexanderJuestel/pyheatdemand#29 @AlexanderJuestel
> # References

The references issues were addressed in https://github.com/AlexanderJuestel/pyheatdemand/commits/main/

I would refrain from adding a reference to the summary as I see it as an abstract and would not add a reference there either.
Two references were added in the statement of need and an example of where to download heat demand input data for the state of North Rhine-Westphalia was addded to "Processing Heat Demand Input Data" Section
