### 2024-02-21 AlexanderJuestel/pyheatdemand#29 @AlexanderJuestel
@nmstreethran with that, I should have addressed your comments and remarks on this issue. Feel free to have a look and thanks again for your great comments!
