### 2024-02-21 #6275 @AlexanderJuestel
> Sorry, I was only halfway through my review. I've opened two more issues now and will complete my checklist once these are resolved. [AlexanderJuestel/pyheatdemand#29](https://github.com/AlexanderJuestel/pyheatdemand/issues/29) [AlexanderJuestel/pyheatdemand#30](https://github.com/AlexanderJuestel/pyheatdemand/issues/30)

@timtroendle, @nmstreethran all comments and remarks in the two referenced issues should be addressed now
