### 2024-02-22 AlexanderJuestel/pyheatdemand#30 @nmstreethran
@AlexanderJuestel I've had a quick look and everything looks good. Many thanks for addressing my comments!
