### 2024-02-22 AlexanderJuestel/pyheatdemand#29 @nmstreethran
> The target audience is included in the Statement of Need: Combining the functionality of well-known geospatial Python libraries, the open-source package PyHeatDemand provides tools for public entities, researchers, or students for processing heat demand input data. Would that be sufficient?

Yes, that's sufficient. I'm happy with all the other changes too.
