+++ pyheatdemand/processing.py
@@ -21,10 +21,12 @@ import geopy
 import osmnx
 
 
-def create_polygon_mask(gdf: Union[gpd.GeoDataFrame, Polygon],
-                        step_size: int,
-                        crop_gdf: bool = False,
-                        crs: Union[str, pyproj.crs.crs.CRS] = None) -> gpd.GeoDataFrame:
+def create_polygon_mask(
+    gdf: Union[gpd.GeoDataFrame, Polygon],
+    step_size: int,
+    crop_gdf: bool = False,
+    crs: Union[str, pyproj.crs.crs.CRS] = None,
+) -> gpd.GeoDataFrame:
     """Create a mask GeoDataFrame consisting of squares with a defined step_size.
 
     Parameters
@@ -75,56 +77,59 @@ def create_polygon_mask(gdf: Union[gpd.GeoDataFrame, Polygon],
     # Converting Polygon to GeoDataFrame
     if isinstance(gdf, Polygon):
         if not isinstance(crs, (str, pyproj.crs.crs.CRS)):
-            raise TypeError('CRS must be provided as string or pyproj CRS object')
-        gdf = gpd.GeoDataFrame(geometry=[gdf],
-                               crs=crs)
+            raise TypeError("CRS must be provided as string or pyproj CRS object")
+        gdf = gpd.GeoDataFrame(geometry=[gdf], crs=crs)
 
     # Checking that the gdf is of type GeoDataFrame
     if not isinstance(gdf, gpd.GeoDataFrame):
-        raise TypeError('gdf must be provided as GeoDataFrame')
+        raise TypeError("gdf must be provided as GeoDataFrame")
 
     # Checking that the step_size is of type int
     if not isinstance(step_size, int):
-        raise TypeError('The step_size must be provided as int')
+        raise TypeError("The step_size must be provided as int")
 
     # Checking that crop_gdf is of type bool
     if not isinstance(crop_gdf, bool):
-        raise TypeError('crop_gdf must be provided as bool')
+        raise TypeError("crop_gdf must be provided as bool")
 
     # Creating arrays
-    x = np.arange(gdf.total_bounds[0],
-                  gdf.total_bounds[2], step_size)
+    x = np.arange(gdf.total_bounds[0], gdf.total_bounds[2], step_size)
 
-    y = np.arange(gdf.total_bounds[1],
-                  gdf.total_bounds[3], step_size)
+    y = np.arange(gdf.total_bounds[1], gdf.total_bounds[3], step_size)
 
     # Creating polygons
-    polygons = [Polygon([(a, b),
-                         (a + step_size, b),
-                         (a + step_size, b + step_size),
-                         (a, b + step_size)]) for a,
-                                                  b in
-                product(x, y)]
+    polygons = [
+        Polygon(
+            [
+                (a, b),
+                (a + step_size, b),
+                (a + step_size, b + step_size),
+                (a, b + step_size),
+            ]
+        )
+        for a, b in product(x, y)
+    ]
 
     # Converting polygons to GeoDataFrame
-    gdf_mask = gpd.GeoDataFrame(geometry=polygons,
-                                crs=gdf.crs)
+    gdf_mask = gpd.GeoDataFrame(geometry=polygons, crs=gdf.crs)
 
     # Dropping duplicate cells
     gdf_mask = gdf_mask.drop_duplicates(ignore_index=True)
 
     # Cropping the gdf if crop_gdf is True
     if crop_gdf:
-        gdf_mask = gdf_mask.sjoin(gdf).reset_index()[['geometry']]
+        gdf_mask = gdf_mask.sjoin(gdf).reset_index()[["geometry"]]
 
     return gdf_mask
 
 
-def refine_mask(mask: gpd.GeoDataFrame,
-                data: gpd.GeoDataFrame,
-                num_of_points: int,
[…]
