### 2024-03-04 #6275 @samuelduchesne
@AlexanderJuestel and @timtroendle I have completed my review. Again, apologies for picking this up so late.

I submitted a few issues which @AlexanderJuestel has already implemented (thank you for that). Overall, this is a good submission; I would consider splitting the `processing.py` into different modules to help with readability and, moving forward, contributions from other developers.
