+++ test/test_processing.py
@@ -7,84 +7,83 @@ from shapely.geometry import Polygon, Point, MultiPolygon, LineString
 import sys
 import os
 
-sys.path.insert(0, '../')
+sys.path.insert(0, "../")
 
 
 def test_create_polygon_mask():
     from pyheatdemand.processing import create_polygon_mask
 
-    gdf = gpd.GeoDataFrame(geometry=[Polygon([(0, 0),  # Bottom-left corner
-                                              (0, 100),  # Top-left corner
-                                              (100, 100),  # Top-right corner
-                                              (100, 0),  # Bottom-right corner
-                                              (0, 0)  # Close the polygon by repeating the first point
-                                              ])],
-                           crs='EPSG:3034'
-                           )
+    gdf = gpd.GeoDataFrame(
+        geometry=[
+            Polygon(
+                [
+                    (0, 0),  # Bottom-left corner
+                    (0, 100),  # Top-left corner
+                    (100, 100),  # Top-right corner
+                    (100, 0),  # Bottom-right corner
+                    (0, 0),  # Close the polygon by repeating the first point
+                ]
+            )
+        ],
+        crs="EPSG:3034",
+    )
 
-    mask = create_polygon_mask(gdf=gdf,
-                               step_size=10,
-                               crop_gdf=False)
-
-    assert isinstance(mask, gpd.GeoDataFrame)
-    assert len(mask) == 100
-    assert all(mask.area == 100) == True
-    assert mask.crs == 'EPSG:3034'
-
-    mask = create_polygon_mask(gdf=gdf,
-                               step_size=10,
-                               crop_gdf=True)
+    mask = create_polygon_mask(gdf=gdf, step_size=10, crop_gdf=False)
 
     assert isinstance(mask, gpd.GeoDataFrame)
     assert len(mask) == 100
     assert all(mask.area == 100) == True
-    assert mask.crs == 'EPSG:3034'
+    assert mask.crs == "EPSG:3034"
 
-    mask = create_polygon_mask(gdf=gdf.iloc[0].geometry,
-                               step_size=10,
-                               crop_gdf=True,
-                               crs='EPSG:3034')
+    mask = create_polygon_mask(gdf=gdf, step_size=10, crop_gdf=True)
 
     assert isinstance(mask, gpd.GeoDataFrame)
     assert len(mask) == 100
     assert all(mask.area == 100) == True
-    assert mask.crs == 'EPSG:3034'
+    assert mask.crs == "EPSG:3034"
+
+    mask = create_polygon_mask(
+        gdf=gdf.iloc[0].geometry, step_size=10, crop_gdf=True, crs="EPSG:3034"
+    )
+
+    assert isinstance(mask, gpd.GeoDataFrame)
+    assert len(mask) == 100
+    assert all(mask.area == 100) == True
+    assert mask.crs == "EPSG:3034"
 
 
 def test_create_polygon_mask_error():
     from pyheatdemand.processing import create_polygon_mask
 
-    gdf = gpd.GeoDataFrame(geometry=[Polygon([(0, 0),  # Bottom-left corner
-                                              (0, 100),  # Top-left corner
-                                              (100, 100),  # Top-right corner
[…]
