### 2024-03-04 #6275 @timtroendle
@AlexanderJuestel , could you please:
- [ ] Remove the superfluous whitespace in the first sentence of your "statement of need" section.
- [ ] Make a tagged release of your software, **and** list the version tag of the archived version here.
- [ ] Archive the reviewed software **in** Zenodo **or** a similar service (e.g., figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., **in** Zenodo) has the correct metadata. This includes the title (should match the paper title) **and** author list (make sure the list **is** correct **and** people who only made a small fix are **not** on it). You may also add the authors’ ORCID.

I can then move forward **with** recommending acceptance of the submission.
