### 2024-03-04 #6275 @AlexanderJuestel
@timtroendle I have removed the white space in https://github.com/AlexanderJuestel/pyheatdemand/commit/22456498e23e3b85f489877e0c331d35c2d52bf0
