+++ CITATION.cff
@@ -1,7 +1,7 @@
-cff-version: 1.2.0
+cff-version: 1.1.0
 message: "If you use PyHeatDemand, please cite it as below."
 authors:
-  - family-names: "Jüstel"
+  - family-names: "Juestel"
     given-names: "Alexander"
     orcid: https://orcid.org/0000-0003-0980-7479
   - family-names: "Strozyk"
@@ -9,7 +9,7 @@ authors:
     orcid: ames: "Alexander"
     orcid: https://orcid.org/0000-0002-3067-831X
 title: "PyHeatDemand - Processing Tool for Heat Demand Data"
-version: 0.0.1
-doi: https://doi.org/10.21105/joss.xxxx
-date-released: 2023-10-31
+version: 0.1.0
+doi: https://doi.org/10.21105/joss.06275
+date-released: 2023-03-04
 url: "https://github.com/AlexanderJuestel/pyheatdemand/"
