+++ CITATION.cff
@@ -6,7 +6,6 @@ authors:
     orcid: https://orcid.org/0000-0003-0980-7479
   - family-names: "Strozyk"
     given-names: "Frank"
-    orcid: ames: "Alexander"
     orcid: https://orcid.org/0000-0002-3067-831X
 title: "PyHeatDemand - Processing Tool for Heat Demand Data"
 version: 0.1.0
