+++ CITATION.cff
@@ -8,7 +8,7 @@ authors:
     given-names: "Frank"
     orcid: https://orcid.org/0000-0002-3067-831X
 title: "PyHeatDemand - Processing Tool for Heat Demand Data"
-version: 0.1.0
+version: 0.1.2
 doi: https://doi.org/10.21105/joss.06275
 date-released: 2023-03-04
 url: "https://github.com/AlexanderJuestel/pyheatdemand/"
