+++ CITATION.cff
@@ -1,7 +1,7 @@
 cff-version: 1.1.0
 message: "If you use PyHeatDemand, please cite it as below."
 authors:
-  - family-names: "Juestel"
+  - family-names: "Jüstel"
     given-names: "Alexander"
     orcid: https://orcid.org/0000-0003-0980-7479
   - family-names: "Strozyk"
