### 2024-03-04 #6275 @AlexanderJuestel
@timtroendle I have created a new release here: https://github.com/AlexanderJuestel/pyheatdemand/releases/tag/v0.1.2

The version is `v0.1.2`

I have also activated the Zenodo connection to the repository. The Repository DOI is  DOI: 10.5281/zenodo.10779452 (https://zenodo.org/records/10779452)

The meta data is correct.
