+++ joss/paper.bib
@@ -198,7 +198,7 @@ author = {Finkel, R.A. and Bentley, J.L. },
 
 @article{Juestel2024_HeatDemand,
     author={Jüstel, A. and Humm, E. and Herbst, E. and Strozyk, F. and Kukla, P. and Bracke, R. },
-    year={2023},
+    year={2024},
     title={Unveiling the spatial distribution of heat demand in {N}orth-{W}est-{E}urope compiled with national heat {C}onsumption data},
     journal={Energies},
     volume = {17},
