### 2024-03-04 #6275 @AlexanderJuestel
I had to change the year of https://doi.org/10.3390/en17020481 from 2023 to 2024. I pushed the change in https://github.com/AlexanderJuestel/pyheatdemand/commit/c3da83c4b9838eb016ea74b71ee210533030d297

I guess that requires a new version and a new archive? I will release a new version and post the new information here.  Sorry for the inconvenience.
