### 2024-03-04 #6275 @AlexanderJuestel
@timtroendle, the new version is `v0.1.3`, the archive is 10.5281/zenodo.10779503
