### 2024-03-05 2d3db10
Updates JOSS paper
