+++ joss/paper.bib
@@ -71,9 +71,9 @@
 }
 
 @software{pandas,
-    author       = {{The pandas development team}},
+    author       = {{Reback,  Jeff and McKinney,  Wes and {Jbrockmendel} and Bossche,  Joris Van Den and Augspurger,  Tom and Cloud,  Phillip and {Gfyoung} and Hawkins,  Simon and {Sinhrks} and Roeschke,  Matthew and Klein,  Adam and {Terji Petersen} and Tratner,  Jeff and She,  Chang and Ayd,  William and Naveh,  Shahar and Garcia,  Marc and {,  Patrick} and Schendel,  Jeremy and Hayden,  Andy and Saxton,  Daniel and Jancauskas,  Vytautas and Shadrach,  Richard and Gorelli,  Marco and McMaster,  Ali and Battiston,  Pietro and {Skipper Seabold} and {Kaiqi Dong} and {chris-b1} and {h-vetinari}}},
     title        = {pandas-dev/pandas: Pandas},
-    month        = march,
+    month        = mar,
     year         = 2021,
     publisher    = {Zenodo},
     version      = {v1.2.3},
@@ -160,7 +160,7 @@ author = {Drilon Meha and Tomislav Novosel and Neven Duić},
 }
 
 @article{herbst,
-title = {A {H}eat {D}emand {M}ap of {N}orth- {W}est {E}urope – its impact
+title = {A {H}eat {D}emand {M}ap of {N}orth-{W}est {E}urope – its impact
 on supply areas and identification of potential production areas for deep geothermal energy},
 journal = {GeoKarlsruhe},
 volume = {2021},
