### 2024-03-05 #6275 @kyleniemeyer
Hi @AlexanderJuestel I made a few small edits, could you review and merge https://github.com/AlexanderJuestel/pyheatdemand/pull/35 ?
