### 2024-03-06 #6275 @kyleniemeyer
Hi @AlexanderJuestel, unfortunately my author change for the pandas citation accidentally included double curly brackets, so the author list appears incorrectly. This PR fixes it: https://github.com/AlexanderJuestel/pyheatdemand/pull/36
