### 2024-03-06 d5e9c4f
Fixes pandas citation in bib (#36)
