### 2024-03-06 #6275 @AlexanderJuestel
@kyleniemeyer @timtroendle thanks for managing the review! It was a pleasure publishing my third paper within the JOSS/JOSE universe! I am glad it went smoothly and I could learn something from the great reviews of @nmstreethran and @samuelduchesne!
